ignore追加＋反映させるためにキャッシュ削除 いらないデータ削除

+++ Assets/Script/SavWav.cs
@@ -14,7 +14,11 @@ public static class SavWav {
 	public static string Initialized(string data)
 	{
 		path = data;
-		Directory.Delete(path,true);
+        if (File.Exists(path))
+        {
+            Directory.Delete(path, true);
+        }
+
 		Directory.CreateDirectory(path);
 		return path;
 	}
